Replace error print stack trace with UncheckedIOException
The methods in ProcessDictionary.java are updated to replace printStackTrace() method calls within IOException catch blocks with a throw UncheckedIOException statement. This provides a cleaner way to deal with exceptions by throwing them up to the calling method, and it better conforms to Java's exception handling best practices.
Also, an extra line was added in DoublingDemo.java to improve code readability. Note that this does not affect the functionality of the code.